+++ fabric/src/main/java/io/github/flemmli97/advancedgolems/fabric/config/ConfigLoader.java
@@ -5,15 +5,16 @@ import io.github.flemmli97.advancedgolems.config.Config;
 public class ConfigLoader {
 
     public static void load() {
-        Config.golemHealth = ConfigSpecs.jsonConfig.getElement().golemHealth;
+        ConfigSpecs specs = ConfigSpecs.get();
+        Config.golemHealth = specs.golemHealth.get();
 
-        Config.golemBaseAttack = ConfigSpecs.jsonConfig.getElement().golemBaseAttack;
+        Config.golemBaseAttack = specs.golemBaseAttack.get();
 
-        Config.homeRadius = ConfigSpecs.jsonConfig.getElement().homeRadius;
-        Config.shouldGearTakeDamage = ConfigSpecs.jsonConfig.getElement().shouldGearTakeDamage;
+        Config.homeRadius = specs.homeRadius.get();
+        Config.shouldGearTakeDamage = specs.shouldGearTakeDamage.get();
 
-        Config.flyItem.readFromString(ConfigSpecs.jsonConfig.getElement().flyItem);
-        Config.speedItem.readFromString(ConfigSpecs.jsonConfig.getElement().speedItem);
-        Config.maxSpeedUpgrades = ConfigSpecs.jsonConfig.getElement().maxSpeedUpgrades;
+        Config.flyItem.readFromString(specs.flyItem.get());
+        Config.speedItem.readFromString(specs.speedItem.get());
+        Config.maxSpeedUpgrades = specs.maxSpeedUpgrades.get();
     }
 }